Decision Tree Calculation

Probabilistically incorporates sampling variability (and standard pathway cost and time to diagnosis).

dectree(data, nsim = 100, costDistns, time_res, drug = drug,
  prevalence = 0.25, performance, c.newtest = 0, name.newtest = NA,
  quant = 0.5, QALYloss, N = nrow(data), wholecohortstats = FALSE,
  terminal_health, terminal_cost, followup_pdf, ...)

Arguments

data
IDEA study data (data.frame)

nsim
Number of sample points (integer) Default: 1000

costDistns
List of distribution names and parameter values for each test/procedure

time_res
time to obtain novel test result (list)

drug
treatment drug cost and quantities (list)

prevalence
TB proportion in cohort (double 0-1) Default: 0.25

performance
Sensitivity and specificity test (list)

c.newtest
Rule-out test unit cost (double) Defalut: 0

name.newtest
Name of rule-out test to get at distribution (string)

quant
Quantile value of time to diagnosis and costs from observed data (double 0-1) Default: 0.5 (median)

QALYloss
QALY loss due to active TB (list)

N
Number of patients. The number in the data is used as default

wholecohortstats
Should the output stats be the total or per patient

terminal_health
function of terminal node total QALY loss

terminal_cost
function of terminal node total costs

followup_pdf
Follow-up appointment time probability distribution function

...

Value

Health and cost realisations (list)

Examples